examples/tapt/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
apex=0.1=pypi_0
argcomplete=3.0.8=pypi_0
assertpy=1.1=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
betty-ml=0.2.1=pypi_0
blas=1.0=mkl
boto=2.49.0=pypi_0
braceexpand=0.1.7=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cached-property=1.5.2=pypi_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
cloudpickle=2.2.1=pypi_0
crcmod=1.7=pypi_0
cryptography=40.0.2=pypi_0
cudatoolkit=11.1.1=h6406543_8
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh4f3eec9_6
datasets=2.0.0=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.4.3=pypi_0
dill=0.3.4=pypi_0
docker-pycreds=0.4.0=pypi_0
fastargs=1.2.0=pypi_0
fasteners=0.18=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.4.1=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=pypi_0
fsspec=2022.1.0=pypi_0
gcs-oauth2-boto-plugin=3.0=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.18=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-apitools=0.5.32=pypi_0
google-auth=2.17.3=pypi_0
google-reauth=0.1.1=pypi_0
gsutil=5.23=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h5py=3.1.0=pypi_0
hdf5=1.10.2=hba1933b_1
httplib2=0.20.4=pypi_0
huggingface-hub=0.4.0=pypi_0
idna=3.3=pypi_0
idna-ssl=1.1.0=pypi_0
imgcat=0.5.0=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
ipdb=0.13.9=pypi_0
ipython=7.16.3=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
learn2learn=0.1.7=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.3.4=pypi_0
mkl=2022.0.1=h06a4308_117
monotonic=1.6=pypi_0
multidict=5.2.0=pypi_0
multiprocess=0.70.12.2=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
ninja=1.10.2.3=pypi_0
nltk=3.6.7=pypi_0
numpy=1.19.5=pypi_0
oauth2client=4.1.3=pypi_0
olefile=0.46=py36_0
opencv=3.4.1=py36h6fd60c2_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1t=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
parso=0.7.1=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.2=py36h06a4308_0
promise=2.3=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pyarrow=6.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.2=pypi_0
pygments=2.12.0=pypi_0
pyopenssl=23.1.1=pypi_0
pyparsing=3.0.7=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pypi_0
pytorch=1.9.0=py3.6_cuda11.1_cudnn8.0.5_0
pytorch-pfn-extras=0.6.6=pypi_0
pytz=2022.1=pypi_0
pyu2f=0.1.5=pypi_0
pyyaml=6.0=pypi_0
qpth=0.0.15=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.3.15=pypi_0
requests=2.27.1=pypi_0
responses=0.17.0=pypi_0
retry-decorator=1.1.1=pypi_0
rsa=4.7.2=pypi_0
sacremoses=0.0.49=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
sentry-sdk=1.5.8=pypi_0
setproctitle=1.2.2=pypi_0
setuptools=58.0.4=py36h06a4308_0
shortuuid=1.0.8=pypi_0
six=1.16.0=pypi_0
sklearn=0.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.38.0=hc218d9a_0
tensorboardx=1.8=pypi_0
termcolor=1.1.0=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pypi_0
torchmetrics=0.8.2=pypi_0
torchvision=0.10.0=py36_cu111
tqdm=4.63.1=pypi_0
traitlets=4.3.3=pypi_0
transformers=4.18.0=pypi_0
triton=1.1.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=pypi_0
wandb=0.12.11=pypi_0
wcwidth=0.2.5=pypi_0
webdataset=0.2.48=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wikiextractor=3.0.6=pypi_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
yaspin=2.1.0=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
